move IOPTranscript to its own file
